parse-util: allow '-0' as alternative to '0' and '+0'
Let's allow "-0" as alternative to "+0" and "0" when parsing integers,
unless the new SAFE_ATO_REFUSE_PLUS_MINUS flag is specified.

In cases where allowing the +/- syntax shall not be allowed
SAFE_ATO_REFUSE_PLUS_MINUS is the right flag to use, but this also means
that -0 as only negative integer that fits into an unsigned value should
be acceptable if the flag is not specified.

CVE: CVE_2020-13776
Upstream-Status: Backport
[https://github.com/systemd/systemd/commit/c78eefc13562a8fc0c22c00a6d3001af89860258]

Signed-off-by: Dan Tran <dantran@microsoft.com>