logind: Restore chvt as non-root user without polkit
4acf0cfd2f ("logind: check PolicyKit before allowing VT switch") broke
the ability to write user sessions that run graphical sessions (e.g.
weston/X11). This was partially amended in 19bb87fbfa ("login: allow
non-console sessions to change vt") by changing the default PolicyKit
policy so that non-root users are again allowed to switch the VT. This
makes the policy when PolKit is not enabled (as on many embedded
systems) match the default PolKit policy and allows launching graphical
sessions as a non-root user.

Closes #17473